backend/requirements.txt
fastapi
uvicorn
temporalio
google-genai
google-adk
resend
supabase
pydantic
pydantic-settings
python-dotenv
httpx
steel-sdk
supermemory
playwright
python-multipart
